requirements.txt
yt-dlp>=2024.08.06
youtube-comment-downloader>=0.1.76
python-dotenv>=1.0.0
pandas>=2.0.0
transformers>=4.35.0
torch>=2.0.0
underthesea>=6.7.0
emoji>=2.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
tqdm>=4.66.0
sentence-transformers>=2.2.0
